fix: time filters
